+++ src/main/kotlin/com/zepben/evolve/streaming/get/NetworkConsumerClient.kt
@@ -621,8 +621,10 @@ class NetworkConsumerClient(
             SWITCHINFO -> extractResult(io.switchInfo.mRID()) { addFromPb(io.switchInfo) }
             CURRENTRELAYINFO -> extractResult(io.currentRelayInfo.mRID()) { addFromPb(io.currentRelayInfo) }
             CURRENTRELAY -> extractResult(io.currentRelay.mRID()) { addFromPb(io.currentRelay) }
-            RECLOSESEQUENCE -> extractResult(io.recloseSequence.mRID()) { addFromPb(io.recloseSequence) }
             OTHER, IDENTIFIEDOBJECT_NOT_SET, null -> throw UnsupportedOperationException(
+                "Identified object type ${io.identifiedObjectCase} is not supported by the network service"
+            )
+            else -> throw UnsupportedOperationException(
                 "Identified object type ${io.identifiedObjectCase} is not supported by the network service"
             )
         }